backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
google-genai>=1.20.0
anthropic>=0.45.0
python-multipart>=0.0.9
httpx>=0.27.0
